Publish Unit Test Results (#1440)
Updated GitHub Actions script
