+++ resilience4j-reactor/src/test/java/io/github/resilience4j/reactor/ratelimiter/operator/MonoRateLimiterTest.java
@@ -68,7 +68,6 @@ public class MonoRateLimiterTest {
 
         StepVerifier.create(
             Mono.error(new IOException("BAM!"))
-                .log()
                 .transformDeferred(RateLimiterOperator.of(rateLimiter)))
             .expectSubscription()
             .expectError(IOException.class)
