+++ resilience4j-spring-boot2/src/test/java/io/github/resilience4j/circuitbreaker/CircuitBreakerAutoConfigurationAsyncTest.java
@@ -15,13 +15,13 @@
  */
 package io.github.resilience4j.circuitbreaker;
 
+import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import io.github.resilience4j.circuitbreaker.autoconfigure.CircuitBreakerProperties;
-import io.github.resilience4j.circuitbreaker.configure.CircuitBreakerAspect;
 import io.github.resilience4j.common.circuitbreaker.monitoring.endpoint.CircuitBreakerEventDTO;
 import io.github.resilience4j.common.circuitbreaker.monitoring.endpoint.CircuitBreakerEventsEndpointResponse;
 import io.github.resilience4j.service.test.DummyService;
-import io.github.resilience4j.service.test.ReactiveDummyService;
 import io.github.resilience4j.service.test.TestApplication;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,6 +32,8 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -39,93 +41,47 @@ import static org.assertj.core.api.Assertions.assertThat;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
     classes = TestApplication.class)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class CircuitBreakerAutoConfigurationRxJava2Test {
+public class CircuitBreakerAutoConfigurationAsyncTest {
 
+    @Rule
+    public WireMockRule wireMockRule = new WireMockRule(8090);
     @Autowired
     CircuitBreakerRegistry circuitBreakerRegistry;
-
     @Autowired
     CircuitBreakerProperties circuitBreakerProperties;
-
-    @Autowired
-    CircuitBreakerAspect circuitBreakerAspect;
-
     @Autowired
     DummyService dummyService;
 
     @Autowired
     private TestRestTemplate restTemplate;
 
-    @Autowired
-    private ReactiveDummyService reactiveDummyService;
-
-
     /**
      * The test verifies that a CircuitBreaker instance is created and configured properly when the
      * DummyService is invoked and that the CircuitBreaker records successful and failed calls.
      */
     @Test
-    public void testCircuitBreakerAutoConfigurationReactiveRxJava2() throws IOException {
+    public void testCircuitBreakerAutoConfigurationAsync()
+        throws IOException, ExecutionException, InterruptedException {
         assertThat(circuitBreakerRegistry).isNotNull();
         assertThat(circuitBreakerProperties).isNotNull();
-
         List<CircuitBreakerEventDTO> circuitBreakerEventsBefore = getCircuitBreakersEvents();
-        List<CircuitBreakerEventDTO> circuitBreakerEventsForBBefore = getCircuitBreakerEvents(ReactiveDummyService.BACKEND);
+        List<CircuitBreakerEventDTO> circuitBreakerEventsForABefore = getCircuitBreakerEvents(DummyService.BACKEND);
 
         try {
-            reactiveDummyService.doSomethingFlowable(true).blockingSubscribe(String::toUpperCase,
-                throwable -> System.out.println("Exception received:" + throwable.getMessage()));
-        } catch (Exception ex) {
-            // Do nothing. The IOException is recorded by the CircuitBreaker as part of the setRecordFailurePredicate as a failure.
-        }
-        // The invocation is recorded by the CircuitBreaker as a success.
-        reactiveDummyService.doSomethingFlowable(false).blockingSubscribe(String::toUpperCase,
-            throwable -> System.out.println("Exception received:" + throwable.getMessage()));
-
-        // expect circuitbreaker-event actuator endpoint recorded both events
-        assertThat(getCircuitBreakersEvents())
-            .hasSize(circuitBreakerEventsBefore.size() + 2);
-        assertThat(getCircuitBreakerEvents(ReactiveDummyService.BACKEND))
-            .hasSize(circuitBreakerEventsForBBefore.size() + 2);
-
-        // Observable test
-        try {
-            reactiveDummyService.doSomethingObservable(true)
-                .blockingSubscribe(String::toUpperCase, Throwable::getCause);
+            dummyService.doSomethingAsync(true);
         } catch (IOException ex) {
             // Do nothing. The IOException is recorded by the CircuitBreaker as part of the setRecordFailurePredicate as a failure.
         }
         // The invocation is recorded by the CircuitBreaker as a success.
-        reactiveDummyService.doSomethingObservable(false)
-            .blockingSubscribe(String::toUpperCase, Throwable::getCause);
+        final CompletableFuture<String> stringCompletionStage = dummyService
+            .doSomethingAsync(false);
+        assertThat(stringCompletionStage.get()).isEqualTo("Test result");
 
-        // Maybe test
-        try {
-            reactiveDummyService.doSomethingMaybe(true).blockingGet("goo");
-        } catch (Exception ex) {
-            // Do nothing. The IOException is recorded by the CircuitBreaker as part of the setRecordFailurePredicate as a failure.
-        }
-        // The invocation is recorded by the CircuitBreaker as a success.
-        reactiveDummyService.doSomethingMaybe(false).blockingGet();
-
-        // single test
-        try {
-            reactiveDummyService.doSomethingSingle(true).blockingGet();
-        } catch (Exception ex) {
-            // Do nothing. The IOException is recorded by the CircuitBreaker as part of the setRecordFailurePredicate as a failure.
-        }
-        // The invocation is recorded by the CircuitBreaker as a success.
-        reactiveDummyService.doSomethingSingle(false).blockingGet();
-
-        // Completable test
-
-        try {
-            reactiveDummyService.doSomethingCompletable(true).blockingAwait();
-        } catch (Exception ex) {
-            // Do nothing. The IOException is recorded by the CircuitBreaker as part of the setRecordFailurePredicate as a failure.
-        }
-        // The invocation is recorded by the CircuitBreaker as a success.
-        reactiveDummyService.doSomethingCompletable(false).blockingAwait();
+        // expect circuitbreaker-event actuator endpoint recorded both events
+        assertThat(getCircuitBreakersEvents())
+            .hasSize(circuitBreakerEventsBefore.size() + 2);
+        assertThat(getCircuitBreakerEvents(DummyService.BACKEND))
+            .hasSize(circuitBreakerEventsForABefore.size() + 2);
     }
 
     private List<CircuitBreakerEventDTO> getCircuitBreakersEvents() {
